+++ apps/client/src/app/pages/i18n/i18n-page.component.ts
@@ -1,13 +1,16 @@
+import { CommonModule } from '@angular/common';
 import { Component, OnInit } from '@angular/core';
 import { Subject } from 'rxjs';
 
 @Component({
   host: { class: 'page' },
+  imports: [CommonModule],
   selector: 'gf-i18n-page',
+  standalone: true,
   styleUrls: ['./i18n-page.scss'],
   templateUrl: './i18n-page.html'
 })
-export class I18nPageComponent implements OnInit {
+export class GfI18nPageComponent implements OnInit {
   private unsubscribeSubject = new Subject<void>();
 
   public constructor() {}